Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -7,6 +7,6 @@ int average(int a,int b,int c){
 int main() {
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    printf("Average: %.2f",average(a,b,c));
+    printf("Average: %.2f",(float)average(a,b,c));
     return 0;
 }
